requirements.txt
python-telegram-bot==20.6
requests
schedule
pandas
pytz
yfinance

python-telegram-bot[ext]==20.6
